backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
openai==1.58.1
anthropic==0.42.0
google-genai==1.1.0
sse-starlette==2.2.1
